refactor: fix PopupNavigator directory name
