+++ src/components/Main/PopupNavigator/PopupNavigator.vue
